requirements.txt
customtkinter~=5.1.3
pillow~=9.5.0
CTkMessagebox~=2.0
matplotlib~=3.7.1
scipy==1.10.1
numpy==1.24.3
